G5 Upgrades:
https://lowendmac.com/2023/maximizing-the-agp-power-mac-g5-into-space-and-beyond/

The G5 tower was a nice machine.  I liked it a
lot.  Other Macs that I regret selling are my
Sawtooth G4, my Quicksilver G4, and my Pismo.

The one Mac that I would like to have owned is
a Lampshade iMac.  ther Macs that might have
been nice to own are a SE/30 and a Quadra 700.