All objects in the Universe:
https://pubs.aip.org/aapt/ajp/article/91/10/819/2911822/All-objects-and-some-questions

"In Fig. 2, we plot all the composite objects in the
Universe: protons, atoms, life forms, asteroids, moons,
planets, stars, galaxies, galaxy clusters, giant voids,
and the Universe itself. Humans are represented by a mass
of 70 kg and a radius of 50 cm (we assume sphericity),
while whales are represented by a mass of 105 kg and a
radius of 7 m."