Starfish are mostly head:
https://www.cnn.com/2023/11/02/world/starfish-head-body-plan-scn/index.html

"[N]ew genetic research suggests ... that sea stars are
largely heads that lack torsos or tails and likely lost
those features evolutionarily over time.  The researchers
said the bizarre fossils of sea star ancestors, which
appeared to have a kind of torso, make a lot more sense in
evolutionary terms in light of the new findings.

...

'When we compared the expression of genes in a starfish
to other groups of animals, like vertebrates, it appeared
that a crucial part of the body plan was missing,' [one
of the researchers] said. 'The genes that are typically
involved in the patterning of the trunk of the animal
weren't expressed in the ectoderm. It seems the whole
echinoderm body plan is roughly equivalent to the head in
other groups of animals.'"