Ancient stars may have made extremely heavy elements:
https://phys.org/news/2023-12-ancient-stars-extraordinarily-heavy-elements.html

"The team took a fresh look at the amounts of heavy
elements in 42 well-studied stars in the Milky Way. ...
By taking a broader view of the amounts of each heavy
element found in these stars collectively, rather than
individually as is more common, they identified previously
unrecognized patterns.

...

Those patterns signaled that some elements listed near the
middle of the periodic table - such as silver and
rhodium - were likely the remnants of heavy element
fission. The team was able to determine that the r-process
can produce atoms with an atomic mass of at least 260
before they fission."