How the Manhattan project was funded:
https://www.nytimes.com/2024/01/17/us/politics/atomic-bomb-secret-funding-congress.html?unlocked_article_code=1.OU0.PuSv.SstGkf8GD8nI

"Where exactly in the budget had the [FDR] administration
hidden the money for the bomb?

Finding a bill that passed eight decades ago is harder
than you might think, but after a number of creative
searches, I pulled up a record of a Senate hearing
examining the bill. Attached was a report breaking down
the legislation.

I paged through, stopped and smiled when I saw it. There
it was, the innocuous phrase that hid an $800 million
secret: 'expediting production.'"