The appendix may not useless:
https://www.npr.org/sections/health-shots/2024/02/02/1228474984/appendix-function-appendicitis-gut-health

"It turns out that the appendix appears to have two
related functions. The first function is supporting the
immune system. The appendix has a high concentration of
immune tissue, so it's acting to help the immune system
fight any bad things in the gut.

The second function that it serves is what we refer to as
the safe house. ... During times of gastrointestinal
distress - you know, a diarrhea episode where all of your
good gut bacteria is getting kind of flushed out of the
system - the appendix is kind of this blind tube with a
very narrow diameter and narrow lumen, so the good
bacteria doesn't get flushed out of the appendix. The idea
is it's safe during this time of gastrointestinal distress
and it can then exit the appendix and recolonize this good
bacteria throughout the rest of the gut."