Protons are complicated:
https://www.quantamagazine.org/inside-the-proton-the-most-complicated-thing-imaginable-20221019/

"Recently, a team led by Juan Rojo of the National
Institute for Subatomic Physics in the Netherlands and VU
University Amsterdam analyzed more than 5,000 proton
snapshots taken over the last 50 years, using machine
learning to infer the motions of quarks and gluons inside
the proton in a way that sidesteps theoretical guesswork.

The new scrutiny picked up a background blur in the images
that had escaped past researchers. In relatively soft
collisions just barely breaking the proton open, most of
the momentum was locked up in the usual three quarks: two
ups and a down. But a small amount of momentum appeared to
come from a 'charm' quark and charm antiquark - colossal
elementary particles that each outweigh the entire proton
by more than one-third."