Tungsten semi-carbine (W2C):
https://phys.org/news/2024-04-dimensional-nanomaterial-expansion.html

"Working at Interface Science Western ... [researchers]
formulated two-dimensional nanosheets of tungsten
semi-carbide (or W2C, a chemical compound containing equal
parts of tungsten and carbon atoms), which, when stretched
in one direction, expand perpendicular to the applied
force. This structural design is known as auxetics.

The trick is that the structure of the nanosheet itself
isn't flat. The atoms in the sheet are made of repeating
units consisting of two tungsten atoms for every carbon
atom, which are arranged metaphorically like the dimpled
surface of an egg carton. As tension is applied across the
elastic nanosheet in one direction, it expands out in the
other dimension as the dimples flatten.

Prior to this innovation, there was only one reported
material that could expand by 10% per unit length in this
counter-intuitive way. The Western-engineered tungsten
semi-carbide nanosheet can expand to 40%, a new world
record."