The forgotten city of Giddan/Eddana (Anqa, Iraq):
https://phys.org/news/2024-04-forgotten-city-identification-dura-europos.html

"The Dura-Europos site in modern-day Syria is famous for
its exceptional state of preservation. ... [T]his
ancient city has yielded many great discoveries, and
serves as a window into the world of the ancient
Hellenistic, Parthian, and Roman periods. ... [T]here is
another city, only some miles down the Euphrates river,
that presents a long-neglected opportunity for study[:] ...
the city of Anqa [] a near mirror image of Dura-Europos,
of the same size, [and] comparable composition ...[.]

...

Anqa is located just across the Syrian border from
Dura-Europos, in the present-day Al-Qaim district of the
Anbar Governorate in Iraq. Its remains include an
identifying tell mound, at the northern end of the site, a
polygonal inner wall circuit, and a large outer defensive
wall, or enceinte. Situated at a point where the Euphrates
floodplain drastically narrows, the city would have
controlled movement between the populous section of the
valley upstream and the trade route downstream linking
Syria, Northern Mesopotamia, and Babylonia, giving it
great strategic and economic significance."