A new instance of primary endosymbiosis:
https://www.independent.co.uk/news/science/algae-evolution-agriculture-plant-history-b2535143.html

"For the first time in at least a billion years, two
lifeforms have merged into a single organism.

The process, called primary endosymbiosis, has only
happened twice in the history of the Earth, with the first
time giving rise to all complex life as we know it through
mitochondria. The second time that it happened saw the
emergence of plants.

Now, an international team of scientists have observed the
evolutionary event happening between a species of algae
commonly found in the ocean and a bacterium.

...

The process involves the algae engulfing the bacterium and
providing it with nutrients, energy and protection in
return for functions that it could not previously perform
- in this instance, the ability to 'fix' nitrogen from
the air.

The algae then incorporates the bacterium as an internal
organ called an organelle, which becomes vital to the
host's ability to function."