How many children did Lady Macbeth have?
https://lifeandletters.substack.com/p/how-many-children-had-lady-macbeth

"We know Lady Macbeth had a child at some point ... [a]nd we
know that Macbeth is concerned about his 'fruitless crown,'
the perennial problem of succession in the absence of an
heir. 'Why does Lady Macbeth allude to a dead infant?' or
'Why are the Macbeths childless?' are perfectly sensible
questions to ask, and many literary scholars have written
very good articles about them. It's the 'how many' that
makes the question ridiculous, demanding a degree of
precision the play will never provide."