Paper types ranked by likelihood of paper cuts:
https://phys.org/news/2024-08-paper-likelihood.html

"The research team found that paper that was the most thin
was unlikely to cause a cut because it tended to buckle
instead. Also, thick paper rarely led to a cut because its
surface was spread over too large an area. That left paper
that is neither too thick nor too thin, like the kind that
is used in newspapers or dot-matrix printers - the two types
tied for the title 'Most likely to cut skin.'

Other culprits were Post-It notes, printed magazines and
office paper.  Some that were less likely to cut include
tissue and photo paper. They noted that to cause cuts, the
paper had to be angled slightly."