A 27-Month Check-In on the Copyright Claims Board:
https://blog.ericgoldman.org/archives/2024/10/a-27-month-check-in-on-the-copyright-claims-board-ccb-guest-blog-post.htm

"Of the 989 claims filed, there are only 819 total
dispositions, meaning 170 claims are still pending. ... Of
the 819 total dispositions, 391 (48%) were dismissed during
claims review, 153 (19%) were not served, 82 (10%) were
withdrawn, and 91 (11%) had defendants opt out.  That means
717 claims (88%) were disposed of on procedural grounds,
without a ruling on the merits. ... In total, the claimant
received a settlement, a default judgment, or a final
judgment in their favor in only about 100 cases, or 12
percent of the 819 dispositions (excluding the cases that
are still pending)."