Yelloway One:
https://phys.org/news/2024-10-banana-resistant-tr4-black-sigatoka.html

"Researchers have developed a banana plant [named Yelloway
One] that is resistant to both fusarium tropical race 4
(TR4) and black sigatoka, two of the most destructive
diseases for bananas.

...

Yelloway One is still a prototype and is currently being
grown in a greenhouse in the Netherlands. The plants will
soon be sent to regions in the Philippines and Indonesia,
where TR4 and black sigatoka have caused significant damage.
Field trials will determine how well Yelloway One performs
in a natural environment."