The CAFC provides additional guidance on inherency:
"The question we must resolve ... is whether a claim limitation
that merely recites an inherent property of an otherwise obvious
combination requires additional analysis to demonstrate that a
person of ordinary skill in the art would have a reasonable
expectation of success. We conclude that this additional showing
is not required."

Cytiva Bioprocess R&D AB v. JSR Corp., No. 23-2075, Slip. Op. at
20 (Fed. Cir. Dec. 4, 2024), available at https://cafc.uscourts.gov/opinions-orders/23-2074.OPINION.12-4-2024_2429911.pdf. [PDF]