3rd Legacy / Outback were peak car:
https://boehs.org/node/old-tech

"This is the 2007 Subaru Legacy 2.5i Special Edition in White. It
just happens to be the pinnacle of the automobile. Sure, cars
have gotten faster. Cars have always been faster. Cars have
gotten smarter. There are cars with better MPG, and there are
cars that are undoubtably a better bang for your buck in today's
market. But there's no car that is as good as this."