Magnetic properties of the quantum Hall effect:
https://phys.org/news/2024-12-current-generated-quantum-hall-effect.html

"The quantum Hall effect, a fundamental effect in quantum
mechanics, not only generates an electric but also a magnetic
current. It arises from the motion of electrons on an orbit
around the nuclei of atoms. This has been demonstrated by the
calculations of a team from Martin Luther University
Halle-Wittenberg[.]"