The physics of random stacking:
https://phys.org/news/2024-12-physics-random-stacking-perspectives-stability.html

"Imagine a tower of kaplas (identical wooden planks) with each
block slightly out of alignment. As the tower rises, the
misalignment increases until an inevitable breaking point is
reached[.] ... This simple phenomenon ... raises a question of
stability: how high can the blocks be stacked before the
structure collapses?

...

The maximum height of a pile before it collapses is inversely
proportional to the square of the amplitude of the positioning
errors. Thus, small errors allow much greater heights to be
reached, while larger errors lead to rapid collapse."