Quipu:
https://www.livescience.com/space/astronomy/astronomers-discover-quipu-the-single-largest-structure-in-the-known-universe

"Astronomers have discovered what may be the largest-scale
structure in the known universe - a group of galaxy clusters and
clusters of galaxy clusters that spans roughly 1.3 billion
light-years across and contains a mind-boggling 200 quadrillion
solar masses.

The newfound structure is dubbed Quipu after an Incan system of
counting and storing numbers using knots on cords.

... [T]he structure is complex, made up of one long filament and
multiple side filaments[,] ...  potentially making it the largest
object in the universe in terms of length, beating out previous
record-holders such as the Laniakea supercluster."