Photographing small birds - focus on the head:
https://www.thephoblographer.com/2025/08/13/the-big-lie-about-bird-photography/

"[F]ocus on the bird's head. This may mean the beak, the
feathers, etc. The depth of field of the lenses that you're using
is often enough to get the whole head in focus. That will
obviously include the eyes too. ...  Focusing on a bigger object
like the head will make keeping the bird in focus a whole lot
easier. Focusing on the neck, depending on the type of bird,
could also work[.]"